refactor utils crate
