+++ http-server/Cargo.toml
@@ -10,12 +10,15 @@ license = "MIT"
 anyhow = "1"
 hyper = { version = "0.14", features = ["stream", "server", "http1", "http2", "tcp"] }
 jsonrpsee-types = { path = "../types", version = "0.2.0-alpha.4" }
-jsonrpsee-utils = { path = "../utils", version = "0.2.0-alpha.4" }
+jsonrpsee-utils = { path = "../utils", version = "0.2.0-alpha.4", features = ["server"] }
+globset = "0.4"
+lazy_static = "1.4"
 log = "0.4"
 serde = { version = "1", default-features = false, features = ["derive"] }
 serde_json = "1"
 socket2 = "0.4"
 tokio = { version = "1", features = ["full"] }
+unicase = "2.6.0"
 
 [dev-dependencies]
 env_logger = "0.8"
