+++ utils/Cargo.toml
@@ -7,22 +7,22 @@ edition = "2018"
 license = "MIT"
 
 [dependencies]
-anyhow = "1"
-futures = "0.3"
-globset = "0.4"
+anyhow = { version = "1", optional = true }
+futures = { version = "0.3", default-features = false, optional = true}
 hyper13 = { package = "hyper", version = "0.13", default-features = false, features = ["stream"], optional = true }
 hyper14 = { package = "hyper", version = "0.14", default-features = false, features = ["stream"], optional = true }
-jsonrpsee-types = { path = "../types", version = "0.2.0-alpha.4" }
-lazy_static = "1.4"
-log = "0.4"
-rustc-hash = "1"
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-serde_json = { version = "1", features = ["raw_value"] }
-tokio = { version = "1", features = ["sync"] }
-unicase = "2.6"
+jsonrpsee-types = { path = "../types", version = "0.2.0-alpha.4", optional = true }
+log = { version = "0.4", optional = true }
+rustc-hash = { version = "1", optional = true }
+serde = { version = "1.0", default-features = false, features = ["derive"], optional = true }
+serde_json = { version = "1", features = ["raw_value"], optional = true }
+tokio = { version = "1", features = ["sync"], optional = true }
 
 [features]
-default = ["hyper14"]
+default = []
+hyper_13 = ["hyper13", "futures", "jsonrpsee-types"]
+hyper_14 = ["hyper14", "futures", "jsonrpsee-types"]
+server = ["anyhow", "tokio", "futures", "jsonrpsee-types", "rustc-hash", "serde", "serde_json", "log"]
 
 [dev-dependencies]
 serde_json = "1.0"
